+++ lib/lseek64.c
@@ -8,7 +8,7 @@
 
 loff_t lseek64(int fildes, loff_t offset, int whence) {
   loff_t tmp;
-  if (llseek(fildes,offset>>32,offset&0xffffffff,&tmp,whence)) {
+  if (llseek(fildes,(unsigned long)(offset>>32),(unsigned long)offset&0xffffffff,&tmp,whence)) {
     if (errno!=ENOSYS) return -1;
     return (loff_t)lseek(fildes,(off_t)offset,whence);
   }
